# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CYFY News Board API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List all workspaces
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspace"
      responses:
        "201":
          description: Created workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspaces/{id}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace by ID with products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Workspace detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspace"
      responses:
        "200":
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteWorkspace
      tags: [workspaces]
      summary: Delete a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Cannot delete master workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspaces/{id}/products:
    post:
      operationId: addProduct
      tags: [workspaces]
      summary: Add a product to workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduct"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspaces/{id}/products/{productId}:
    delete:
      operationId: removeProduct
      tags: [workspaces]
      summary: Remove a product from workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /workspaces/{id}/feed:
    get:
      operationId: getWorkspaceFeed
      tags: [workspaces]
      summary: Get threat feed for a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Workspace threat feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceFeedResponse"
  /workspaces/{id}/match:
    post:
      operationId: matchWorkspaceThreats
      tags: [workspaces]
      summary: Trigger threat matching for workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Matching result
          content:
            application/json:
              schema:
                type: object
                required:
                  - matchedCount
                properties:
                  matchedCount:
                    type: integer
  /workspaces/{id}/matches/{matchId}:
    put:
      operationId: updateMatch
      tags: [workspaces]
      summary: Update a threat match (review/resolve/dismiss)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: matchId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reviewed:
                  type: boolean
                dismissed:
                  type: boolean
                matchStatus:
                  type: string
                  enum: [active, resolved]
      responses:
        "200":
          description: Updated match
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ThreatIntelItem:
      type: object
      required:
        - id
        - title
        - summary
        - description
        - scope
        - severity
        - category
        - source
        - ttps
        - iocs
        - malwareFamilies
        - affectedSystems
        - mitigations
        - targetSectors
        - targetRegions
        - status
        - confidenceLevel
        - publishedAt
        - updatedAt
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        description:
          type: string
        scope:
          type: string
          enum: [local, global]
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        category:
          type: string
        threatActor:
          type: string
          nullable: true
        threatActorAliases:
          type: array
          items:
            type: string
        targetSectors:
          type: array
          items:
            type: string
        targetRegions:
          type: array
          items:
            type: string
        ttps:
          type: array
          items:
            type: string
        iocs:
          type: array
          items:
            type: string
        malwareFamilies:
          type: array
          items:
            type: string
        affectedSystems:
          type: array
          items:
            type: string
        mitigations:
          type: array
          items:
            type: string
        source:
          type: string
        sourceUrl:
          type: string
          nullable: true
        references:
          type: array
          items:
            type: string
        campaignName:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, resolved, monitoring]
        confidenceLevel:
          type: string
          enum: [confirmed, high, medium, low]
        firstSeen:
          type: string
          nullable: true
        lastSeen:
          type: string
          nullable: true
        publishedAt:
          type: string
        updatedAt:
          type: string
        isIndiaRelated:
          type: boolean
          nullable: true
        indiaConfidence:
          type: integer
          nullable: true
        indianState:
          type: string
          nullable: true
        indianStateName:
          type: string
          nullable: true
        indianCity:
          type: string
          nullable: true
        indianSector:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
    Workspace:
      type: object
      required:
        - id
        - name
        - domain
        - isDefault
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        domain:
          type: string
        description:
          type: string
          nullable: true
        isDefault:
          type: boolean
        createdAt:
          type: string
          nullable: true
        updatedAt:
          type: string
          nullable: true
    WorkspaceDetail:
      allOf:
        - $ref: "#/components/schemas/Workspace"
        - type: object
          properties:
            products:
              type: array
              items:
                $ref: "#/components/schemas/Product"
    CreateWorkspace:
      type: object
      required:
        - name
        - domain
      properties:
        name:
          type: string
        domain:
          type: string
        description:
          type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/CreateProduct"
    UpdateWorkspace:
      type: object
      properties:
        name:
          type: string
        domain:
          type: string
        description:
          type: string
    Product:
      type: object
      required:
        - id
        - productName
        - enabled
      properties:
        id:
          type: string
          format: uuid
        productName:
          type: string
        vendor:
          type: string
          nullable: true
        version:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        enabled:
          type: boolean
    CreateProduct:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        vendor:
          type: string
        version:
          type: string
        category:
          type: string
    WorkspaceFeedResponse:
      type: object
      required:
        - items
        - total
        - page
        - limit
      properties:
        items:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/ThreatIntelItem"
              - type: object
                properties:
                  matchId:
                    type: string
                    format: uuid
                  matchedProduct:
                    type: string
                  relevanceScore:
                    type: number
                  reviewed:
                    type: boolean
                  matchStatus:
                    type: string
                    enum: [active, resolved]
                  resolvedSeverity:
                    type: string
                    enum: [critical, high, medium, low, info]
                    nullable: true
                  resolvedAt:
                    type: string
                    nullable: true
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer